+++ Okta.AspNetCore/OpenIdConnectOptionsHelper.cs
@@ -44,6 +44,7 @@ namespace Okta.AspNetCore
                 "okta-aspnetcore",
                 typeof(OktaAuthenticationOptionsExtensions).Assembly.GetName().Version,
                 oktaMvcOptions);
+            oidcOptions.BackchannelTimeout = oktaMvcOptions.BackchannelTimeout;
 
             var hasDefinedScopes = oktaMvcOptions.Scope?.Any() ?? false;
             if (hasDefinedScopes)
@@ -99,6 +100,7 @@ namespace Okta.AspNetCore
             jwtBearerOptions.Events = oktaWebApiOptions.JwtBearerEvents ?? new JwtBearerEvents();
             jwtBearerOptions.SecurityTokenValidators.Clear();
             jwtBearerOptions.SecurityTokenValidators.Add(new StrictSecurityTokenValidator());
+            jwtBearerOptions.BackchannelTimeout = oktaWebApiOptions.BackchannelTimeout;
         }
     }
 }